Functions

rel_path_d

DynamicHMCModels.rel_path_d — Method.

relpathd

Relative path using the DynamicHMCModels src/ directory.

Example to get access to the data subdirectory

rel_path_d("..", "data")

source

generate_d

DynamicHMCModels.generate_d — Method.

generate_d

Utility function to generate all notebooks and chapters from scripts in the scripts directory.

Method

generate_d(sd = script_dict_d)

Required arguments

None, all notebooks/.. and chapters/.. files are regenerated.

source

DynamicHMCModels.generate_d — Method.

generate

Generate notebooks and scripts in a single chapter.

Method

generate(chapter::AbstractString)

Required arguments

Generate notebooks and scripts in a single chapter, e.g. generate("04")

source

DynamicHMCModels.generate_d — Method.

generate

Generate a single notebook and script

Method

generate(chapter::AbstractString, file::AbstractString)

Required arguments

Generate notebook and script file in chapter, e.g. generate("04", "m4.1d.jl") or generate("04/m4.1d.jl")

source